Copy more (plotting related) stuff from legacy
