+++ reflexible/utils.py
@@ -4,6 +4,7 @@ from datetime import datetime
 
 from matplotlib.dates import date2num
 import numpy as np
+import xarray as xr
 
 
 class Structure(dict):
@@ -49,3 +50,83 @@ def closest(num, numlist):
         numlist = date2num(numlist)
 
     return (np.abs(numlist - num)).argmin()
+
+
+def data_range(data, min='median'):
+    """
+    return a data range for flexpart data
+
+    optional keyword min = ['median', 'mean', 'min']
+    """
+    if isinstance(data, xr.Variable):
+        data = data.values
+
+    dmax = np.nanmax(data)
+    if np.isnan(dmax):
+        dmax = 1e5
+
+    data = data[data.nonzero()]
+    if min == 'mean':
+        dmin = np.mean(data)
+    elif min == 'median':
+        dmin = np.median(data)
+    else:
+        dmin = np.nanmin(data)
+
+    if np.isnan(dmin):
+        dmin = 1e-5
+
+    return [dmin, dmax]
+
+
+#
+# Curtain related functions
+#
+
+def __groundlevel_for_line(H, X, Y, coords, index=0):
+    """
+    extracts ground level from H.heightnn along a track of lon, lat
+
+    input:  H or H.heightnn (takes the lowest level)
+
+            X, Y ,= H.longitude, H.latitude
+
+            coords = zip(x, y)
+
+    output: groundlevel (a 1-d array with shape (len(flighttrack)
+
+    """
+    try:
+        hgt = H.Heightnn[:, :, 0]
+    except Exception:
+        hgt = H
+
+    # fix for hgt offset
+    hgt = hgt - hgt.min()
+
+    grndlvl = np.zeros(len(coords))
+    for i, (x, y) in enumerate(coords):
+        I = closest(x, X)
+        J = closest(y, Y)
+        grndlvl[i] = hgt[I, J]
+        grndlvl = np.nan_to_num(grndlvl)
+
+    return grndlvl - grndlvl.min()
+
+
+def curtain_agltoasl(H, curtain_agl, coords, below_gl=0.0):
+    """ converts the agl curtain to asl
+
+        adds .asl_axis attribute to H
+    """
+
+    gl = __groundlevel_for_line(H, H.longitude, H.latitude, coords)
+    H.asl_axis = np.linspace(0, H.outheight[-1])
+    xp = H.outheight - H.outheight[0]
+    casl = np.zeros((len(H.asl_axis), len(coords)))
+
+    for i in range(len(coords)):
+        casl[:, i] = np.interp(H.asl_axis, xp + gl[i], \
+                              curtain_agl[:, i], left=below_gl)
+
+    return casl
